backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
selenium>=4.20,<5.0
openai>=1.30,<3.0
chromadb>=0.5,<2.0
sentence-transformers>=2.7,<6.0
numpy>=1.26,<3.0
pymysql>=1.1,<2.0
